+++ packages/wiz-ui-next/src/components/base/snackbar/snackbar.stories.ts
@@ -22,21 +22,36 @@ export default {
         type: "text",
       },
     },
+    expand: {
+      control: {
+        type: "boolean",
+      },
+    },
   },
   parameters: {
     layout: "fullscreen",
   },
 } as Meta<typeof WizSnackbar>;
 
-export const Snackbar: StoryFn<typeof WizSnackbar> = (args) => ({
+const Template: StoryFn<typeof WizSnackbar> = (args) => ({
   components: { WizSnackbar },
   setup() {
     return { args };
   },
-  template: `<WizSnackbar :message="args.message" :isStatic="true" />`,
+  template: `<WizSnackbar v-bind="$props"/>`,
 });
-Snackbar.args = {
+
+export const Default = Template.bind({});
+Default.args = {
   message: "Hello World Hello World Hello World Hello World",
+  isStatic: true,
+};
+
+export const Expand = Template.bind({});
+Expand.args = {
+  message: "Hello World Hello World Hello World Hello World",
+  isStatic: true,
+  expand: true,
 };
 
 export const UseSnackbar: StoryFn<typeof WizSnackbar> = () => ({
@@ -63,3 +78,22 @@ UseSnackbar.parameters = {
     },
   },
 };
+
+export const UseSnackbarWithExpand: StoryFn<typeof WizSnackbar> = () => ({
+  setup() {
+    const message = ref("Hello World");
+    const options = ref({ expand: true });
+    const { snack } = useSnackbar();
+    return {
+      message,
+      options,
+      snack,
+    };
+  },
+  template: `
+    <div style="text-align: center;">
+      <input v-model="message" />
+      <button @click="snack(message, options)">Show</button>
+    </div>
+  `,
+});