+++ packages/wiz-ui-react/src/components/base/calendar/components/calendar-helper.ts
@@ -2,7 +2,8 @@ import { CalendarData, CalendarDataItem } from "./types";
 
 export function createCalendarData(
   currentShowingDateTime: Date,
-  shouldFillWeeks?: boolean
+  shouldFillWeeks?: boolean,
+  disabledDate?: (date: Date) => boolean
 ): CalendarData {
   const year = currentShowingDateTime.getFullYear();
   const month = currentShowingDateTime.getMonth();
@@ -24,6 +25,9 @@ export function createCalendarData(
     length: firstDayOfMonth,
   })
     .map((_, index) => ({
+      isDisabled: !!disabledDate?.(
+        new Date(year, month - 1, LastDateOfLastMonth - index)
+      ),
       label: shouldFillWeeks ? String(LastDateOfLastMonth - index) : "",
       isOutOfCurrentMonth: true,
     }))
@@ -33,6 +37,7 @@ export function createCalendarData(
   const currentMonthDataItems: CalendarDataItem[] = Array.from({
     length: lastDateOfMonth,
   }).map((_, index) => ({
+    isDisabled: !!disabledDate?.(new Date(year, month, index + 1)),
     label: String(index + 1),
   }));
 
@@ -40,6 +45,7 @@ export function createCalendarData(
   const nextMonthDataItems: CalendarDataItem[] = Array.from({
     length: 6 - lastDayOfMonth,
   }).map((_, index) => ({
+    isDisabled: !!disabledDate?.(new Date(year, month + 1, index + 1)),
     label: shouldFillWeeks ? String(index + 1) : "",
     isOutOfCurrentMonth: true,
   }));