+++ packages/wiz-ui-react/src/components/icons/remove-bold.tsx
@@ -1,0 +1,10 @@
+export const WizIRemoveBold = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    height="1em"
+    viewBox="0 -960 960 960"
+    width="1em"
+  >
+    <path d="M240-440q-17 0-28.5-11.5T200-480q0-17 11.5-28.5T240-520h480q17 0 28.5 11.5T760-480q0 17-11.5 28.5T720-440H240Z" />
+  </svg>
+);